requirements.txt
numpy
torch
transformers
gradio
python-docx
pdfplumber
rank-bm25

